Restore tests running in ./util module
Also move testutil.Must* from lakefs/pkg/testutil to lakefs/util/testutil.  We
will probably want to break up lakefs/pkg/testutil, it is a fairly random
assembly of utilities used in testing.
