+++ pkg/config/config_test.go
@@ -17,8 +17,8 @@ import (
 	"github.com/treeverse/lakefs/pkg/block/local"
 	"github.com/treeverse/lakefs/pkg/config"
 	"github.com/treeverse/lakefs/pkg/kv/kvparams"
-	"github.com/treeverse/lakefs/pkg/testutil"
 	"github.com/treeverse/lakefs/util/logging"
+	"github.com/treeverse/lakefs/util/testutil"
 )
 
 func newConfigFromFile(fn string) (*config.Config, error) {
