+++ pkg/graveler/committed/meta_range_writer_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/treeverse/lakefs/pkg/graveler"
 	"github.com/treeverse/lakefs/pkg/graveler/committed"
 	"github.com/treeverse/lakefs/pkg/graveler/committed/mock"
-	"github.com/treeverse/lakefs/pkg/testutil"
+	"github.com/treeverse/lakefs/util/testutil"
 )
 
 func getExpected(t *testing.T, record graveler.ValueRecord) committed.Record {
