+++ pkg/graveler/ref/commit_ordered_iterator_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/treeverse/lakefs/pkg/graveler"
 	"github.com/treeverse/lakefs/pkg/graveler/ref"
 	"github.com/treeverse/lakefs/pkg/kv/mock"
-	"github.com/treeverse/lakefs/pkg/testutil"
+	"github.com/treeverse/lakefs/util/testutil"
 )
 
 func newFakeAddressProvider(commits []*graveler.Commit) *fakeAddressProvider {
