+++ pkg/graveler/ref/link_addresss_iterator_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/go-test/deep"
 	"github.com/treeverse/lakefs/pkg/graveler"
 	"github.com/treeverse/lakefs/pkg/graveler/ref"
-	"github.com/treeverse/lakefs/pkg/testutil"
+	"github.com/treeverse/lakefs/util/testutil"
 )
 
 func TestLinkAddressIterator(t *testing.T) {
