+++ pkg/graveler/ref/merge_base_finder_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/treeverse/lakefs/pkg/graveler"
 	"github.com/treeverse/lakefs/pkg/graveler/ref"
-	"github.com/treeverse/lakefs/pkg/testutil"
+	"github.com/treeverse/lakefs/util/testutil"
 )
 
 type MockCommitGetter struct {
