+++ util/go.work
@@ -1,1 +1,3 @@
 go 1.21.3
+
+use .
